+++ snowflake/table_snowflake_database_grant.go
@@ -23,7 +23,7 @@ func tableSnowflakeDatabaseGrant(_ context.Context) *plugin.Table {
 				{Name: "database"},
 			},
 		},
-		Columns: []*plugin.Column{
+		Columns: snowflakeColumns([]*plugin.Column{
 			{Name: "database", Type: proto.ColumnType_STRING, Transform: transform.FromField("Name").Transform(valueFromNullable), Description: "Name of the database."},
 			{Name: "privilege", Type: proto.ColumnType_STRING, Description: "A defined level of access to an database."},
 			{Name: "created_on", Type: proto.ColumnType_TIMESTAMP, Description: "Date and time when the access was granted."},
@@ -32,7 +32,7 @@ func tableSnowflakeDatabaseGrant(_ context.Context) *plugin.Table {
 			{Name: "granted_on", Type: proto.ColumnType_STRING, Description: "Type of the object."},
 			{Name: "granted_to", Type: proto.ColumnType_STRING, Description: "Type of the object role has been granted."},
 			{Name: "grantee_name", Type: proto.ColumnType_STRING, Description: "Name of the object role has been granted."},
-		},
+		}),
 	}
 }