+++ snowflake/table_snowflake_network_policy.go
@@ -21,7 +21,7 @@ func tableSnowflakeNetworkPolicy(_ context.Context) *plugin.Table {
 		List: &plugin.ListConfig{
 			Hydrate: listSnowflakeNetworkPolicies,
 		},
-		Columns: []*plugin.Column{
+		Columns: snowflakeColumns([]*plugin.Column{
 			{Name: "name", Type: proto.ColumnType_STRING, Description: "Identifier for the network policy."},
 			{Name: "created_on", Type: proto.ColumnType_TIMESTAMP, Description: "Date and time when the policy was created."},
 			{Name: "comment", Type: proto.ColumnType_STRING, Description: "Specifies a comment for the network policy."},
@@ -29,7 +29,7 @@ func tableSnowflakeNetworkPolicy(_ context.Context) *plugin.Table {
 			{Name: "entries_in_blocked_ip_list", Type: proto.ColumnType_INT, Description: "No of entries in the blocked IP list."},
 			{Name: "allowed_ip_list", Type: proto.ColumnType_STRING, Hydrate: DescribeNetworkPolicy, Transform: transform.FromField("ALLOWED_IP_LIST"), Description: "Comma-separated list of one or more IPv4 addresses that are allowed access to your Snowflake account."},
 			{Name: "blocked_ip_list", Type: proto.ColumnType_STRING, Hydrate: DescribeNetworkPolicy, Transform: transform.FromField("BLOCKED_IP_LIST"), Description: "Comma-separated list of one or more IPv4 addresses that are denied access to your Snowflake account."},
-		},
+		}),
 	}
 }